fix(USB-CDC): prevent write buffer clear during log download
Root cause: _usb_write_fail_count > 100 (~100ms) threshold was too
aggressive for large log downloads. When USB endpoint saturated during
sustained LOG_DATA transfer, the write buffer was cleared, dropping
in-flight MAVLink packets. The GCS never received them, causing the
download to stall at ~95% (35MB/37MB).

Fixes:
- Increase USB CDC tx buffer from 2048 to 8192 bytes to absorb bursts
- Raise _usb_write_fail_count threshold from 100 to 5000 (5s)
- Change _usb_write_fail_count type from uint8_t to uint16_t for range

+++ libraries/AP_HAL_RTT/UARTDriver.h
@@ -73,7 +73,7 @@ private:
     bool _initialized;
     bool _deferred_open;
     bool _is_usb{false};
-    uint8_t _usb_write_fail_count{0};
+    uint16_t _usb_write_fail_count{0};
     enum flow_control _flow_control = FLOW_CONTROL_DISABLE;
     uint16_t _last_options = 0;
     uint32_t _tx_stats_bytes = 0;
